Move gist shortcode into its own JS file for better organization
