+++ .eleventy.js
@@ -28,6 +28,58 @@ module.exports = function(eleventyConfig) {
   eleventyConfig.addWatchTarget("./css/");
   eleventyConfig.addWatchTarget("./js/");
 
+    // Customize Markdown library and settings:
+  let markdownLibrary = markdownIt({
+      html: true,
+      linkify: true,
+      typographer: true
+    }).use(markdownItAnchor, {
+      permalink: markdownItAnchor.permalink.headerLink(),
+      level: [1,2,3,4],
+      slugify: eleventyConfig.getFilter("slugify")
+    });
+
+    // Wrap images in a figure, a, and figcaption.
+    // This lets the simplelightbox code serve it up too!
+    // Also adds loading lazy attribute
+    markdownLibrary.renderer.rules.image = function (tokens, idx, options, env, slf) {
+
+      const token = tokens[idx];
+      // Set the loading=lazy attribute
+      token.attrSet('loading', 'lazy');
+
+      // Adjust the path so it works with the pathPrefix
+      // This can be / or /my-blog for example
+      let imgPath = token.attrGet('src');
+      if(!imgPath.includes(pathPrefix) && !imgPath.includes('://')){
+        imgPath = path.join(pathPrefix, imgPath);
+      }
+      token.attrSet('src', imgPath);
+
+      let captionRendered = markdownLibrary.renderInline(token.content);
+      let figCaption = '';
+
+      // If it's a gallery of images, display the caption in the lightbox.
+      // Otherwise display it on the page inside figcaption.
+      // This is because the caption might be too long and awkward to display
+      // in a crowded area.
+      if(env.inGallery){
+        token.attrSet('title', captionRendered);
+      }
+      else {
+        figCaption = captionRendered;
+      }
+
+      // Return figure with figcaption.
+      // The 'a' is the image linking to itself, which then gets picked up by simplelightbox
+      return `<figure><a href="${token.attrs[token.attrIndex('src')][1]}">
+      ${slf.renderToken(tokens, idx, options)}</a>
+      <figcaption>${figCaption}</figcaption>
+    </figure>`;
+    }
+
+    eleventyConfig.setLibrary("md", markdownLibrary);
+
   // Add plugins
   eleventyConfig.addPlugin(pluginRss);
   eleventyConfig.addPlugin(pluginSyntaxHighlight);
@@ -157,65 +209,9 @@ module.exports = function(eleventyConfig) {
   // Show the current year using a shortcode
   eleventyConfig.addShortcode("year", () => `${new Date().getFullYear()}`);
 
-  // Shortcode to fetch and display a gist
-  eleventyConfig.addNunjucksAsyncShortcode("gist", async (gistId) => {
-      let url = `https://api.github.com/gists/${gistId}`;
-      let fetchOptions = {};
-
-      let gistJson = {};
-      let mdCode = '';
-
-      // If we're running in a Github Action, the token can be used to make Github API calls with better rate limits.
-      // Otherwise, without this, sometimes the API call fails due to a low rate limit.
-      // To pass the token, npm run build --ghtoken=${{ secrets.GITHUB_TOKEN }}
-      if (process.env.npm_config_ghtoken){
-        console.log('in the if block')
-        fetchOptions.headers = { 'Authorization': `Bearer ${process.env.npm_config_ghtoken}`}
-      }
-
-      /* fetch() returns a promise, but await can be used inside addNunjucksAsyncShortcode */
-      let response = await fetch(url, fetchOptions);
-
-      if(response.ok){
-        gistJson = await response.json();
-      }
-      else {
-        console.log(await response.json());
-        return '';
-      }
-
-
-      let description = gistJson.description;
-
-      for (var f in gistJson.files){
-        let language = gistJson.files[f].language;
-        let content = gistJson.files[f].content
-        let fileName = gistJson.files[f].filename;
-
-        if(!language){
-          language = "text"
-        }
-        if(language==="C#"){
-          language = "csharp";
-        }
-
-        language = language.toLowerCase();
-
-
-        if(language==="markdown"){
-          //Special case for MD, just render it outright.
-          mdCode += `\n${content}\n`;
-        }
-        else {
-          //Surround by code tags so it gets rendered with prism
-          mdCode += `\`${fileName}\`\n\`\`\`${language}\n${content}\n\`\`\`\n`;
-        }
-
-      }
-
-      return `**${description}**\n${markdownLibrary.render(mdCode)}`;
-
-  });
+  var GistShortCode = require('./_configs/gist.shortcode');
+  let gist = new GistShortCode(markdownLibrary);
+  eleventyConfig.addNunjucksAsyncShortcode("gist", async(gistId) => { return await gist.getGist(gistId); } );
 
   // Create an array of all tags
   eleventyConfig.addCollection("tagList", function(collection) {
@@ -227,57 +223,7 @@ module.exports = function(eleventyConfig) {
     return filterTagList([...tagSet]);
   });
 
-  // Customize Markdown library and settings:
-  let markdownLibrary = markdownIt({
-    html: true,
-    linkify: true,
-    typographer: true
-  }).use(markdownItAnchor, {
-    permalink: markdownItAnchor.permalink.headerLink(),
-    level: [1,2,3,4],
-    slugify: eleventyConfig.getFilter("slugify")
-  });
 
-  // Wrap images in a figure, a, and figcaption.
-  // This lets the simplelightbox code serve it up too!
-  // Also adds loading lazy attribute
-  markdownLibrary.renderer.rules.image = function (tokens, idx, options, env, slf) {
-
-    const token = tokens[idx];
-    // Set the loading=lazy attribute
-    token.attrSet('loading', 'lazy');
-
-    // Adjust the path so it works with the pathPrefix
-    // This can be / or /my-blog for example
-    let imgPath = token.attrGet('src');
-    if(!imgPath.includes(pathPrefix) && !imgPath.includes('://')){
-      imgPath = path.join(pathPrefix, imgPath);
-    }
-    token.attrSet('src', imgPath);
-
-    let captionRendered = markdownLibrary.renderInline(token.content);
-    let figCaption = '';
-
-    // If it's a gallery of images, display the caption in the lightbox.
-    // Otherwise display it on the page inside figcaption.
-    // This is because the caption might be too long and awkward to display
-    // in a crowded area.
-    if(env.inGallery){
-      token.attrSet('title', captionRendered);
-    }
-    else {
-      figCaption = captionRendered;
-    }
-
-    // Return figure with figcaption.
-    // The 'a' is the image linking to itself, which then gets picked up by simplelightbox
-    return `<figure><a href="${token.attrs[token.attrIndex('src')][1]}">
-    ${slf.renderToken(tokens, idx, options)}</a>
-    <figcaption>${figCaption}</figcaption>
-  </figure>`;
-  }
-
-  eleventyConfig.setLibrary("md", markdownLibrary);
 
   // Override Browsersync defaults (used only with --serve)
   eleventyConfig.setBrowserSyncConfig({
